fix: changed packages for DTO to match Core branch
