+++ src/main/java/com/wwme/wwme/task/domain/DTO/receiveDTO/CreateTagReceiveDTO.java
@@ -1,4 +1,4 @@
-package com.wwme.wwme.task.domain.DTO.tagReceiveDTO;
+package com.wwme.wwme.task.domain.DTO.receiveDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class UpdateTagReceiveDTO {
-    private Long tag_id;
+public class CreateTagReceiveDTO {
+    private long group_id;
     private String tag_name;
 }
