+++ src/main/java/com/wwme/wwme/task/domain/DTO/sendDTO/ReadTaskListByUserSendDTO.java
@@ -1,4 +1,4 @@
-package com.wwme.wwme.task.domain.DTO.taskSendDTO;
+package com.wwme.wwme.task.domain.DTO.sendDTO;
 
 import lombok.*;
