+++ src/main/java/com/wwme/wwme/task/domain/DTO/sendDTO/TaskListReadbyUserSendDTO.java
@@ -1,4 +1,4 @@
-package com.wwme.wwme.task.domain.DTO.taskReceiveDTO;
+package com.wwme.wwme.task.domain.DTO.sendDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -9,14 +9,16 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class UpdateTaskReceiveDTO {
+public class TaskListReadbyUserSendDTO {
     private Long task_id;
     private String task_name;
     private LocalDateTime start_time;
     private LocalDateTime end_time;
     private String task_type;
     private Long tag_id;
+    private String tag_name;
     private Long group_id;
-    private Boolean is_done_total;
+    private String group_color;
     private Boolean is_done_personal;
+    private Boolean is_done_total;
 }
