+++ src/main/java/com/wwme/wwme/task/service/TagCRUDServiceImpl.java
@@ -2,14 +2,12 @@ package com.wwme.wwme.task.service;
 
 import com.wwme.wwme.group.domain.Group;
 import com.wwme.wwme.group.repository.GroupRepository;
-import com.wwme.wwme.task.domain.DTO.TagDTO;
-import com.wwme.wwme.task.domain.DTO.tagReceiveDTO.CreateTagReceiveDTO;
-import com.wwme.wwme.task.domain.DTO.tagReceiveDTO.UpdateTagReceiveDTO;
-import com.wwme.wwme.task.domain.DTO.taskSendDTO.TagListReadSendDTO;
+import com.wwme.wwme.task.domain.DTO.receiveDTO.CreateTagReceiveDTO;
+import com.wwme.wwme.task.domain.DTO.receiveDTO.UpdateTagReceiveDTO;
+import com.wwme.wwme.task.domain.DTO.sendDTO.TagListReadSendDTO;
 import com.wwme.wwme.task.domain.Tag;
 import com.wwme.wwme.task.repository.TagRepository;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
